refactor : context/provider 구조 리팩토링
