+++ service/src/App.tsx
@@ -3,7 +3,7 @@ import Header from './components/common/Header/Header';
 import EventInfoFooter from './components/common/Footer/EventInfoFooter';
 import Footer from './components/common/Footer/Footer';
 import Tooltip from './components/common/Tooltip/Tooltip';
-import { TabProvider } from './store/context/TabContext';
+import { TabProvider } from './store/provider/TabProvider';
 
 const App = () => {
   return (
